backend/requirements.txt
Django>=5.0
django-unfold==0.51.0
psycopg2-binary==2.9.9
redis==5.1.1
django-redis==5.4.0
httpx==0.27.2
stripe==10.12.0
Pillow==10.4.0
boto3==1.35.36
django-storages==1.14.4
python-dotenv==1.0.1
openpyxl==3.1.5
gunicorn==23.0.0
whitenoise==6.7.0
